+++ CHANGELOG.md
@@ -248,9 +248,9 @@ We can do that with a macro that finds the respective sub-navigation, and sets a
 
 ```js
 up.macro('.main-nav a', function(link, { section }) {
-  let subLink = document.querySelectorAll(`.sub-nav[data-section="${section}"]`)
-  let subURLs = up.util.map(subLink, 'href')
-  link.setAttribute(subURLs.join())
+  let subLinks = document.querySelectorAll(`.sub-nav[data-section="${section}"]`)
+  let subURLs = up.util.map(subLinks, 'href')
+  link.setAttribute('up-alias', subURLs.join())
 })
 ```